Unit Tests  * Added checkout_donation_confirm tests
Helpers
 * Added error handling to the login helper.

Client
 * Added the ability to use http or https protocol when needed for a local environment.

Provider
 * Added a protocol attribute that will take on the apiendpointProtocol or if nothing specified it will default to https
