+++ lib/client/client.js
@@ -100,6 +100,7 @@ function Client(options) {
     // Semi-common
     this.apiEndpoint = null;
     this.apiEndpointPort = null;
+    this.apiEndpointProtocol = options.protocol || null;
     this.userToken = null;
 
     // Flags and settings
@@ -152,6 +153,7 @@ Client.prototype.isProviderRegistered = function(name) {
 };
 
 
+//noinspection JSUnusedGlobalSymbols
 /**
  * Registers a new API provider with the client
  * @param {string} name – Name of the provider
