+++ lib/provider/http_provider.js
@@ -178,7 +178,8 @@ HttpProvider.prototype._executeRequest = function(headers, query, path, method, 
     this._httpRequest({
         options: requestOptions,
         startTime: startTime,
-        data: data
+        data: data,
+        protocol: this.api.apiEndpointProtocol || https
     }, callback);
 };
